Start working on Ch2 of AAS
I'm trying to do this work in gradle, rather than maven. Maven
is just unusable from the shell.

My work corresponds to page 33 of the text - thus far I don't have
a class with a main for batch jobs, but the build works and defines a
utility class I can use from the spark shell.

I include in the readme the commands to (a) build and (b) run an
interactive shell with the built jar included.