Find a good stopping point for Chapter 3
Unfortunately, it looks like with this dataset and the computer
I am running on (a single node system with 6 Gb ram), ALS on this
dataset runs out of heap memory.

It's likely I could run on a bigger instance, tune my jvm, and/or
get a real cluster up and running. But in the interests of making quick
progress I am going to just read the rest of the chapter carefully
and move on. I at least finished the data extraction code.
  (I'm pretty sure the jvm is limiting me b/c my instance is supposed to
   have 6 GB of ram. Although it's possible the cloud I'm running on
   has overprovisioned and that's why I'm running out of ram)