Start ISLR
Start making Rmd files for chapters of "An Introduction to Statistical
Learning". So far I have a quick exploration of datasets
which demonstrates some plotting methods, plus a lab from Chapter 2
that demonstrates loading data, a few more plots, some data
manipulation, and a basic run of KNN for classification and regression.