Wrote some short notes on Chapter 5
I did some work on doing the Chapter 5 code, but
ultimately I found the author's explanations too
lacking to really follow along. I found it frustrating
so I gave up; I'll finish Chapter 4 and call that good
enough for this book, so I can move on to other resources
which hopefully teach more idiomatic D3 (my opinion is that
"Getting Started..." covers the mechanics ok but the idioms
not so well).